# docs/openapi.yml
openapi: 3.0.3
info:
  title: h24s_10
  version: 0.0.1
  description: traP 24 spring hackathon team 10

paths:
  /ping:
    get:
      description: Ping endpoint
      tags:
        - health
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: pong
  /events:
    summary: event
    description: event description
    post:
      description: create a new event
      tags:
        - event
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEventRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostEventResponse"
  /events/me:
    get:
      tags:
        - event
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventMeResponses"
  /events/{eventID}:
    parameters:
      - $ref: "#/components/parameters/eventID"
    get:
      tags:
        - event
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEventResponse"
  /events/{eventID}/confirm:
    parameters:
      - $ref: "#/components/parameters/eventID"
    patch:
      tags:
        - event
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchEventConfirmRequest"
      responses:
        204:
          description: OK
  /events/{eventID}/participants:
    parameters:
      - $ref: "#/components/parameters/eventID"
    get:
      description: イベントの確定した参加者を取得
      tags:
        - event
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEventParticipantsResponse"
  /events/{eventID}/targets:
    parameters:
      - $ref: "#/components/parameters/eventID"
    get:
      description: イベントの参加候補者を取得
      tags:
        - event
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEventTargetsResponse"
  /events/{eventID}/applicants:
    parameters:
      - $ref: "#/components/parameters/eventID"
    post:
      description: 都合のいい日程候補を提出
      tags:
        - event
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEventApplicantsRequest"
      responses:
        201:
          description: Created
    get:
      description: イベント参加希望を出した人たちを取得
      tags:
        - event
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEventApplicantsResponse"
  /traq/groups:
    get:
      description: get all traQ groups
      tags:
        - traq
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTraQGroupsResponse"
  /traq/users:
    get:
      description: get all traQ users
      tags:
        - traq
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTraQUsersResponse"

components:
  schemas:
    PostEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dateOptions:
          type: array
          items:
            type: object
            properties:
              start:
                type: string
                format: date-time
              end:
                type: string
                format: date-time
            required:
              - start
              - end
        targets:
          type: array
          items:
            type: string
            description: traQ ID
      required:
        - title
        - description
        - dateOptions
        - targets
    PostEventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
      required:
        - id
    EventMeResponse:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        isConfirmed:
          type: boolean
        isAnswered:
          type: boolean
        isHost:
          type: boolean
      required:
        - event_id
        - title
        - isConfirmed
        - isAnswered
        - isHost
    EventMeResponses:
      type: array
      items:
        $ref: "#/components/schemas/EventMeResponse"
    GetEventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        isConfirmed:
          type: boolean
        date:
          $ref: "#/components/schemas/DateTimeResponse"
        dateOptions:
          type: array
          items:
            $ref: "#/components/schemas/dateOption"
      required:
        - id
        - title
        - description
        - isConfirmed
        - dateOptions
    DateTimeResponse:
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
      required:
        - start
        - end
    PatchEventConfirmRequest:
      type: object
      properties:
        isConfirmed:
          type: boolean
        eventDateOptionID:
          type: string
          format: uuid
      required:
        - isConfirmed
    PostEventApplicantsRequest:
      type: object
      properties:
        dateOptionIDs:
          type: array
          items:
            type: string
            format: uuid
            description: 日時候補のID
    GetEventParticipantsResponse:
      type: array
      items:
        type: string
        description: traQ user ID
    GetEventTargetsResponse:
      type: array
      items:
        type: string
        description: traQ userID
    GetEventApplicantsResponse:
      type: array
      items:
        $ref: "#/components/schemas/applicant"
    GetTraQGroupsResponse:
      type: array
      items:
        $ref: "#/components/schemas/traQGroup"
    GetTraQUsersResponse:
      type: array
      items:
        $ref: "#/components/schemas/traQUser"
    dateOption:
      type: object
      properties:
        id:
          type: string
          format: uuid
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
      required:
        - id
        - start
        - end
    applicant:
      type: object
      properties:
        traqID:
          type: string
        dateOptionIDs:
          type: array
          items:
            type: string
            format: uuid
    traQUser:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
      required:
        - name
        - displayName
    traQGroup:
      type: object
      properties:
        name:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/traQUser"

  parameters:
    eventID:
      name: eventID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    groupName:
      name: groupName
      in: path
      required: true
      schema:
        type: string